com.unity.playablegraph-visualizer@0.1.0-preview.1 # Changelog
All notable changes to this package will be documented in this file.

The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).

## [0.1.0] - 2018-12-20

### First release of Unity package *PlayableGraph Visualizer*.

Mirrored the original [graph-visualizer repository](https://github.com/Unity-Technologies/graph-visualizer)
and modified it to fulfill the Unity package requirements.